Move API serializers to their own module.
